refactor: Ventiladores renders from the block registry
fanRpm/temps/curve become self-contained blocks over a ref-counted useFanState
singleton (one shared monitor poll). The temps block declares itself unavailable
in the solo fan+temp case, matching the merged-card layout.
