+++ src/fans/logic.ts
@@ -1,5 +1,18 @@
 // Pure helpers for the fan monitor: a rolling sample buffer and an SVG sparkline
 // path. Kept pure so they're unit-testable; components consume them.
+import type { FanState } from "../api";
+
+/** "Solo" cooling: exactly one fan + one temp — a single system, shown merged
+ *  (fan ring + its temperature) in one card instead of two blocks. */
+export function isSolo(state: FanState): boolean {
+  return state.fans.length === 1 && state.temps.length === 1;
+}
+
+/** Whether the standalone temperatures block has content: the device reports
+ *  temps AND it isn't the solo case (where the single temp lives in the fan card). */
+export function tempsAvailable(state: FanState | null): boolean {
+  return !!state && !isSolo(state) && state.temps.length > 0;
+}
 
 /** Append `value`, keeping at most `max` most-recent samples. Returns a new array. */
 export function pushSample(buffer: number[], value: number, max: number): number[] {
