+++ src/fans/useFanState.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useSyncExternalStore } from "react";
 import { getFanState, FanState } from "../api";
 import { pushSample } from "./logic";
 
@@ -14,44 +14,56 @@ export interface FanMonitor {
 const EMPTY: FanMonitor = { state: null, fanHistory: {} };
 
 /**
- * Polls get_fan_state() every ~1.5 s while mounted (i.e. only while the
- * Ventiladores tab is open) and keeps a rolling RPM buffer per fan for the
- * sparklines. History is keyed by label, not array index, so a fan keeps its
- * own buffer even when curation changes the set (e.g. the generic acpi_fan is
- * dropped once a vendor chip appears) — no cross-contaminated graphs. One state
- * update per tick (single render). Never throws.
+ * Ref-counted singleton monitor: one get_fan_state() poll (~1.5 s) shared by every
+ * consumer, so the fanRpm/temps/curve blocks each read the same tick instead of
+ * registering their own poll. History is keyed by label, not array index, so a fan
+ * keeps its own buffer even when curation changes the set (no cross-contaminated
+ * graphs). One snapshot update per tick (single render). Never throws.
  */
-export function useFanState(): FanMonitor {
-  const [monitor, setMonitor] = useState<FanMonitor>(EMPTY);
+class FanStateStore {
+  private monitor: FanMonitor = EMPTY;
+  private refs = 0;
+  private timer: ReturnType<typeof setInterval> | null = null;
+  private listeners = new Set<() => void>();
 
-  useEffect(() => {
-    let alive = true;
-    const tick = () => {
-      getFanState()
-        .then((s) => {
-          if (!alive) return;
-          setMonitor((prev) => {
-            const fanHistory: Record<string, number[]> = {};
-            for (const f of s.fans) {
-              const prevBuf = prev.fanHistory[f.label] ?? [];
-              // Skip glitch reads (rpm null) so they never poison the sparkline.
-              fanHistory[f.label] =
-                f.rpm == null ? prevBuf : pushSample(prevBuf, f.rpm, MAX_SAMPLES);
-            }
-            return { state: s, fanHistory };
-          });
-        })
-        .catch(() => {
-          /* keep last values */
-        });
-    };
-    tick();
-    const timer = setInterval(tick, POLL_MS);
+  private tick = (): void => {
+    getFanState()
+      .then((s) => {
+        const prev = this.monitor;
+        const fanHistory: Record<string, number[]> = {};
+        for (const f of s.fans) {
+          const prevBuf = prev.fanHistory[f.label] ?? [];
+          // Skip glitch reads (rpm null) so they never poison the sparkline.
+          fanHistory[f.label] = f.rpm == null ? prevBuf : pushSample(prevBuf, f.rpm, MAX_SAMPLES);
+        }
+        this.monitor = { state: s, fanHistory };
+        this.listeners.forEach((l) => l());
+      })
+      .catch(() => {
+        /* keep last values */
+      });
+  };
+
+  subscribe = (cb: () => void): (() => void) => {
+    this.listeners.add(cb);
+    if (this.refs++ === 0) {
+      this.tick();
+      this.timer = setInterval(this.tick, POLL_MS);
+    }
     return () => {
-      alive = false;
-      clearInterval(timer);
+      this.listeners.delete(cb);
+      if (--this.refs === 0 && this.timer) {
+        clearInterval(this.timer);
+        this.timer = null;
+      }
     };
-  }, []);
+  };
 
-  return monitor;
+  getSnapshot = (): FanMonitor => this.monitor;
+}
+
+const store = new FanStateStore();
+
+export function useFanState(): FanMonitor {
+  return useSyncExternalStore(store.subscribe, store.getSnapshot, store.getSnapshot);
 }
